Add post-processor for llvm-sgx to encrypt enclave
The post-processor encrypts everything in the enclave except
__llvmsgx_enclave_entry (which contains the decrypter), and
__llvmsgx_enclave_data (which contains encrypted regions and the key).

The encryption used is XTEA in with the CTR block cipher mode of operation.
This is used mainly because it is very easy to implement in the enclave
without needing to keep track of a lot of functions that must not be
encrypted by the post-processor.